chore: move icons directory
