+++ src/application/components/icons/type.ts
